refactor: 더보기 설정 버튼에서 테마 모달 관련 컴포넌트 분리
